requirements.txt
numpy
matplotlib
scikit-learn
pandas
streamlit
plotly-express
colormap
graphviz
seaborn
xgboost
openpyxl